[MOV] website_sale: split classes into separate files
