+++ addons/website_sale/models/website_visitor.py
@@ -3,12 +3,6 @@
 from odoo import api, fields, models
 
 
-class WebsiteTrack(models.Model):
-    _inherit = 'website.track'
-
-    product_id = fields.Many2one('product.product', ondelete='cascade', readonly=True, index='btree_not_null')
-
-
 class WebsiteVisitor(models.Model):
     _inherit = 'website.visitor'
